Add filtering tests and fix bug in query_for_has_many
If a parent class contains multiple has_many entries of that all use the
same predicate the `query_for_has_many` would still generate the full
list of reflections instead of only returning reflections from the
filtered for children. Adding a "has_model_ssim" check to the query when
filtering is occurring prevents the errant results.